feat(flow-graph): refactor and optimize LabelNode components for improved tooltip functionality
